[FIX] run LowerGvInitializers before ShadowDsa
seahorn assumes that static global variables are explicitly initialized.
This can be done right before ShadowDsa pass since it gives an
opportunity for other passes to remove global variables.

Need to check whether initialization can be a complex constant
expression. We might be assuming that those are lowered. In which case a
call to lower such expressions is necessary after LowerGvInitializers.