Finished array bounds check analysis. usage: sea abc test.bc        sea clang-abc test.c
The analysis instruments each memory access with an assertion that is
satisfied iff the memory access is in-bounds. The analysis splits the
original program into N instrumented programs and runs each one in
parallel, where N is the number of heap and stack allocation
sites. The original program is safe iff the N instrumented programs
are safe. Each of the N instrumented programs instruments only the
memory accesses confined in the corresponding allocation site.

The analysis can also analyze all checks without splitting by
allocation site by using option --do-not-split-proof.