[IMPROVED] Covered the instrumentation of some missed allocations in ABC2:
In ABC2, when a pointer can be allocated we add instrumentation that
deterministically chooses whether we keep track of it or not. There
were several missed cases where we were instrumenting a memory access
without having instrumented its allocation:

- AllocaInst / Malloc-like with unknown size.
- IntToPtrInst
- Return value of external calls
- InsertValueInst/ExtractValueInst.

The above cases are now covered except
for InsertValueInst/ExtractValueInst where we print a warning instead.