Extended instrumentation of load/store in ABC2 and ABC3
[ABC3] added two functions to be optionally filled out:
- sea_abc_log_load_ptr (int8_t *lhs, int8t *ptr)
  add extra assumptions about the pointer to be read.
- sea_abc_log_load_ptr (int8_t *lhs, int8t *ptr)
  add extra assertions about the pointer to be written.

Right now, only abc4.c fills out those functions following the below
description.

[ABC2] If --abc-local-reasoning=true then add some extra
instrumentation for load/stores.

For load instructions, assume about the pointer read from memory that
if it is valid then it is equal to the base pointer being tracked. For
store instructions, assert that if the pointer to be stored is valid
then it is equal to the base.